+++ frontend/src/pages/Browse/DataTable.tsx
