+++ frontend/src/pages/Browse/ResultCard.tsx
