+++ frontend/src/routes/routes.tsx
@@ -6,7 +6,7 @@ import { Simulation } from '@/App';
 import Compare from '@/pages/Compare/Compare';
 import Docs from '@/pages/Docs/Docs';
 import Home from '@/pages/Home/Home';
-import Search from '@/pages/Search/Search';
+import Search from '@/pages/Browse/Browse';
 import Upload from '@/pages/Upload/Upload';
 
 interface RoutesProps {
@@ -28,7 +28,7 @@ const createRoutes = ({
       element: <Home simulations={simulations} />,
     },
     {
-      path: '/search',
+      path: '/Browse',
       element: (
         <Search
           simulations={simulations}
